refactor: rename the drizzle module to become the db module
