+++ src/app.module.ts
@@ -1,11 +1,11 @@
 import { Module } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { DrizzleModule } from './drizzle/drizzle.module';
+import { DrizzleModule } from './db/db.module';
 import { ConfigModule } from '@nestjs/config';
 import { UsersController } from './users/users.controller';
 import { UsersModule } from './users/users.module';
-import dbConfig from './drizzle/config/dbConfig';
+import dbConfig from './db/config/dbConfig';
 import { drizzleConfig } from 'drizzle.config';
 
 @Module({
